refactor(tests): consolidate pure guardrail gates into tests/_guardrails/ (stages 2+3)
Stage 2: move eight pure architecture/doc gates from tests/unit/ to
tests/_guardrails/ via git mv (history preserved):
test_cli_boundary, test_type_boundaries, test_docs_module_refs,
test_install_docs, test_middleware_context_contract, test_cassette_shapes,
test_docs_examples_idioms, test_public_surface.

test_docstrings.py stays in tests/unit/ — it is a hybrid (the two
parametrized docstring-drift checks are gates, but test_from_storage_smoke_
constructs_client is a behavioral end-to-end client smoke test), so it is
not a pure gate.

Stage 3: split the two hybrids — gate functions move to a new tests/_guardrails/
file, behavioral functions stay in tests/unit/:

* test_public_shims.py -> tests/_guardrails/test_public_surface_manifest.py
  (38 manifest / re-export-identity / frozen-__all__ / removed-name /
  facade-delegates / auth-seam-manifest gates). The 10 behavioral functions
  (research/ResearchAPI back-compat delegations, UnknownTypeWarning filter +
  parser-dedup state, auth cookie-validation behaviour, rpc_call forwarding)
  stay in test_public_shims.py.

* test_init_order.py -> tests/_guardrails/test_no_facade_reach_in.py
  (26 AST reach-in / runtime-import boundary gates + their visitor self-tests
  + module-deletion asserts, plus the shared AST visitor classes/helpers). The
  14 construction / init-order behaviour tests (constructor-DI seams, seam
  wiring, phase tests, sources-before-notebooks wiring, mind-map flows) stay in
  test_init_order.py and import the shared AST helpers from the guardrails file.

Update every reference to the moved/split files across docs, ADRs, and test
helpers; fix the cross-module import in test_cassette_patterns.py to the
package-qualified _guardrails.test_cassette_shapes path. Refresh the
Boundary Guardrails and Architecture-gates sections in docs/development.md.
