+++ tests/_guardrails/test_cassette_shapes.py
