+++ tests/_guardrails/test_docs_examples_idioms.py
