+++ tests/_guardrails/test_middleware_context_contract.py
